keyfile: updated connection when writing keyfile
Writing a connection to keyfile and read it back should result
in the identical setting. That is, a full round-trip would not
alter any information.

That is however particularly not true for certificate properties, where
the keyfile writes blobs to file and coerces paths.

Thus, whenver writing a keyfile we must read back what we just
wrote and use that instead.