ifcfg/trivial: rename svFileSetName() and svFileSetModified() to mark for testing
The filename of the shvarFile instance should be immutable and stay
unchanged for the entire lifetime of the instance. Similarly, the
modified flag should not be explicitly set.

However, for testing it is still useful to give the unit test a
direct access to those functions.

Rename the setters to make it clear that this is test-only.