manager: don't schedule devices_inited_cb() a second time
The devices_inited_cb() callback is really supposed to only run
when there is nothing else left in the mainloop to dispatch.

But as we already schedule the idle action with G_PRIORITY_LOW+10
priority, it is very unlikely that there is anything else ready
to run (unless scheduled with an even lower priority, and then it
wouldn't help either because devices_inited_cb() would win again).